个人中心 添加

+++ cunqianguan/Common/ThirdPart/PopoverView/PopoverView.m
@@ -795,7 +795,7 @@
 
 - (void)didTapButton:(UIButton *)sender
 {
-    int index = [subviewsArray indexOfObject:sender];
+    NSInteger index = [subviewsArray indexOfObject:sender];
     
     if (index == NSNotFound) {
         return;
